# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Monitor",
        "contact": {
            "name": "Glassen_IT"
        },
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://62.113.103.54/api",
            "description": "API Monitor"
        }
    ],
    "paths": {
...
        "/pages": {
            "get": {
                "tags": [
                    "pages"
                ],
                "summary": "Список страниц",
                "operationId": "be285396fe3a2c5171e59e42660abcc8",
                "parameters": [
                    {
                        "name": "site_id",
                        "in": "query",
                        "description": "Id сайта(если не указывать, то прийдет всё)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int"
                        }
                    },
                    {
                        "name": "start(0)",
                        "in": "query",
                        "description": "Отсчет для пагинации",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int"
                        }
                    },
                    {
                        "name": "limit(20)",
                        "in": "query",
                        "description": "Лимит пагинации",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer"
                                                    },
                                                    "url": {
                                                        "type": "string"
                                                    },
                                                    "size": {
                                                        "type": "integer"
                                                    }
                                                },
                                                "type": "object"
                                            },
                                            "example": [
                                                {
                                                    "id": 637347,
                                                    "url": "/",
                                                    "size": 1323
                                                },
                                                {
                                                    "id": "...",
                                                    "url": "...",
                                                    "size": "..."
                                                }
                                            ]
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Не авторизован",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResult401"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/pages/{page_id}": {
            "delete": {
                "tags": [
                    "pages"
                ],
                "summary": "Удаленеи страницы",
                "operationId": "5db3d1b3831ad101d2fc0c442e89d55e",
                "responses": {
                    "204": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Не авторизован",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResult401"
                                }
                            }
                        }
                    }
                }
            },
            "patch": {
                "tags": [
                    "pages"
                ],
                "summary": "Редактирование страницы",
                "operationId": "5caa69e2760c1bb3cbdc4ec9bbf91daa",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "url": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "url": "/newurl"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "204": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Не авторизован",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResult401"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ErrorResult401": {
                "title": "Шаблон ошибки статуса авторизации",
                "properties": {
                    "message": {
                        "type": "string"
                    }
                },
                "type": "object",
                "example": {
                    "message": "Unauthenticated."
                }
            }
        }
    },
...
}